Revert "Merge pull request #30 from jamesgoodhouse/add_envfromsecret_for_datasources_sidecar"
This reverts commit f4cd46e1a3994d7616208ad0c34eba0f8b57aa74, reversing
changes made to 31bc9e299058e62af11eb1a78089e856705d44fe.

This was an unnecessary change to allow envFromSecret to be used for the
sidecar for setting up datasources. It is not needed there, but instead
in the main grafana container that will consume the defined datasources,
and that functionality already exists for the main container.

Signed-off-by: James J. Goodhouse <j.goodhouse@gmail.com>