[promtail] Fix scrape configs
The default config was missing a few drop actions. This had caused logs
to be added twice, once the correct way and once with an incorrect job
label.

In order to help debug the config, the flag
`config.snippets.addScrapeJobLabel` is introduced. If enabled, an
additional label `scrape_job` is added which identifies the scrape job
that picked up the log message.

Signed-off-by: Reinhard Nägele <unguiculus@gmail.com>