[promtail] Disable initContainer by default
The Promtail initContainer is enabled by default which sets the
`fs.inotify.max_user_instances` sysctl to `128`.
This sysctl is not namespaced and it's value is applied to the node.
Also the default value of `128` is too low and can cause issues for other pods
running on bigger nodes.
Therefore the sysctl must be disabled by default.

Signed-off-by: Siavash Safi <siavash.safi@gmail.com>