[promtail] Tolerate node-role.kubernetes.io/control-plane:NoSchedule taints
The Kubernetes project is moving away from offensive wording and as part
of this KEP-2067 was accepted to replace the
`node-role.kubernetes.io/master:NoSchedule` taint with
`node-role.kubernetes.io/control-plane:NoSchedule`.

To be compatible with clusters that already use the new taint this pull
request adds an additional toleration for it. This change can't really
break existing setups because we continue to tolerate the old taint as
well.

Additional links:
- https://github.com/kubernetes/enhancements/blob/master/keps/sig-cluster-lifecycle/kubeadm/2067-rename-master-label-taint/README.md
- https://github.com/kubernetes/enhancements/issues/2067
- https://github.com/kubernetes/kubeadm/issues/2200

Signed-off-by: Max Rosin <git@hackrid.de>