[grafana] Fix default for automounting of SA token
In #571 / b6015e2 an option was introduced to optionally automount the
serviceaccount token. Due to an oversight at my part, I missed the fact
that a lot of people (including myself) use side cars for
dashboards/datasource discovery from K8S configmaps. So the default
should be to enable the automount of the SA token, to keep these
sidecars working.

Signed-off-by: Jorik Jonker <jorik.jonker@eu.equinix.com>