[grafana] Update README.md
This commit updates the Grafana Helm chart documentation.

In the section about the image renderer plugin, the reference was linking to
a broken documentation in another GitHub repository.

After inspecting the changes on the `grafana-image-render` repository,
I found out that the change below updated most of the documentation,
where the content of the `remote_rendering_using_docker.md` was
moved to the main `README.md` file.

Then, this fix points towards the new reference.

Here it is the commit that changed the documentation on the remote repo.

https://github.com/grafana/grafana-image-renderer/pull/277

Signed-off-by: Willian Braga <willian.braga@sybogames.com>