Bump chart image to r168-eb3f021e instead of v1.6.1
There are a few changes that are available in the r168 release needed to
make Graphite work.

These changes can't be backported yet to the official semver. This will
most likely happen with GEM v2.0.

Because of this the image tag of the helm chart is now pointing to r168
instead of the semver release v1.6.1.

Also made some changes to the config arguments because multiple flags
have been modified and they are required to make every service work.

Signed-off-by: Jesus Vazquez <jesus.vazquez@grafana.com>