Allow rendering mimir templates in other contexts
When `mimir-distributed` is used as a subchart, the rendering of the
annotation "checksum/config" on Pods is wrong. This is important to
trigger service restart when the config is changed.

There are two options:

a) config is in `mimir-distributed` or shared with `mimir-distributed` in
which case it can contain templates and values of the parent chart, making
it impossible to evaluate in the subchart.

b) config lives in the parent chart, but then it is not possible to reach
it in `mimir-distributed` as subchart and the checksum is the checksum of
an empty string. See: "few important details" in
[Helm subcharts and globals](https://helm.sh/docs/chart_template_guide/subcharts_and_globals/)

This patch makes the manifest templates global named templates and thus
able to be rendered in the parent context. It is one step short of
merging the mimir-distributed chart with enterprise-metrics.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>