Fix whitespace issue due to empty value
If not using the ingress feature the domain value would be empty. This
resulted in the template producing an ini file with a line like this:

`domain = \n`

This would cause issues when passing the resource through a tool like
kustomize, which would turn the multiline block into an encoded single
line string in order to preserve trailing whitespace, resulting in
kubectl diffs that are very hard to review.

Signed-off-by: Gustaf Lindstedt <gustaf.lindstedt@embark-studios.com>