[agent-operator] Add container SecurityContext (#1829)
This is useful for when running in a restricted environment. For example when running in a namespace using the Restricted policy (either via a PodSecurityPolicy or via the new namespace labels), the agent-operator container should not be allowed privilege escalation and should only have access to a read-only root filesystem.

These are already configurable in the CRDs for the Grafana Agent pods that the operator creates, but the operator doesn't have them.