[grafana] fix image-renderer network policy
The currently defined network policy only works when the image renderer
is matching the authorized namespace based on the `name` label, which is
not a label all kubernetes namespace have, but instead a widely used but
nevertheless not universal label.

The default label for all namespace, added by the API server, is
`kubernetes.io/metadata.name`, as documented here:
https://kubernetes.io/docs/reference/labels-annotations-taints/#kubernetes-io-metadata-name

By changing this labelSelector for the `namespaceSelector` clause, we
permit users of kubernetes distribution where no `name` label is
automatically set on namespace to use the image renderer without issue,
and we don't impact the usability of other users.

Signed-off-by: Clément Nussbaumer <clement@astutus.org>