Update CODEOWNERS
Here we make sure that each chart has a Grafana team owner, as well as
extra community members.  This will allow a change to be made to the
repo shortly after merge that will enforce two approvals per PR and
avoid a situation where a contributor can circumvent the controls by
using a separate account.

Signed-off-by: Zach Leslie <zach.leslie@grafana.com>