fix(grafana): Fixed image rendered network policy namespace selector
The egress network policy for the image rendered expects the namespace
of the chart to match `name: ${value}`, this label is not applied to
namespaces by default in Kubernetes and would need to be added to the
namespace in order for this network policy to work. However the ingress
network policy uses `kubernetes.io/metadata.name: ${value}`, this label
is applied to namespaces in kubernetes by default and will allow the
network policy to work out of the box with reasonable defaults.

Without this patch, essentially grafana is able to communicate with the
image renderer, but the image renderer is not able to make HTTP requests
back to grafana in the same namespace without adding `name: ${namespace
name}` to the labels of the namespace grafana is deployed in.

Signed-off-by: Elliot Courant <me@elliotcourant.dev>