Add default emptyDir for /tmp with synthetic-agent
The synthetic-monitoring-agent needs a writeable `/tmp` to allow k6
script to be run. When running as a non root user and a read only fs
this will break currently.

This PR adds a default emptyDir for the /tmp folder so this will work.
I did not add optional volume mounts here since i do not see the value
in that for this pod.

Signed-off-by: BlackDex <black.dex@gmail.com>