[tempo-distributed] Add serviceName to the metrics-generator statefulset
## Problem Description

When running helm validation it is expected that each statefulset contains a serviceName as an identifier [which is known to cause issues with validation](https://github.com/helm/helm/issues/7207). Other statefulset components such as ingesters have this field and thus pass validation. 

## PR Description

This change simply adds `serviceName: metricsGenerator` to the spec field of the metrics generator statefulset template. 

All other component statefulsets in tempo-distributed have a serviceName which meets this requirement, this change brings them in line. 

Signed-off-by: sblack79 <sjb4634@gmail.com>