refactor(tempo-distributed): include servicemonitor for (jaeger) tempo-query only if enabled
Currently the service monitor is added regardless of whether the old jaeger mode is enabled or not.
Similar to the ports on the deployment and service, we only include the servicemonitor if `queryFrontend.query.enabled=true`.
Additionally, the selector of the service monitor does not work anyway, because the service that includes the old ports has a different `app.kubernetes.io/component` label than the one created from the servicemonitor template

Signed-off-by: Alexander Bartolomey <github@alexanderbartolomey.de>
Signed-off-by: Martin Burgos <lqk@qlik.com>